requirements.txt
Flask==3.0.0
requests==2.31.0
Pillow==10.2.0
numpy>=1.26.0
python-dotenv==1.0.0
huggingface_hub>=0.20.0
